Changed filenames to be more suitable.
The "html.txt" was changed to "dump.txt" to reflect that it is the "dumpfile" of "creoledump" that the HTML will by default be dumped into.

The "creole" folder was taken from python-creole-1.3.2, but it was only a part of the python-creole module, so it didn't include the whole project, and it had even left out the AUTHORS and LICENSE files - all because I hadn't figured out how to import modules from subfolders. beyond one level deep. Now it includes the entire python-creole-1.3.2 module in the python-creole-1.3.2/ folder and links to the creole module inside it from the creoledump script.

It also finally includes the LICENSE file, linking to GPL v3. Same as the python-creole project used.